+++ adc.h
@@ -1,3 +1,3 @@
 void adc_setup(void);
 void adc_start(uint8_t adc_ch);
-void adc_read(float *vswr_val);
+void adc_read(uint16_t *vswr_val);
